# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blossom Buddy API",
        "description": "API pour la gestion des plantes et utilisateurs dans l'application Blossom Buddy",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/plants": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Liste toutes les plantes",
                "description": "Récupérer la liste complète de toutes les plantes disponibles",
                "operationId": "6e2a3a91967b1e364c87b755cd5152f6",
                "responses": {
                    "200": {
                        "description": "Liste des plantes récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "common_name": {
                                                "type": "string",
                                                "example": "Rose"
                                            },
                                            "watering_general_benchmark": {
                                                "properties": {
                                                    "value": {
                                                        "type": "string",
                                                        "example": "7"
                                                    },
                                                    "unit": {
                                                        "type": "string",
                                                        "example": "days"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Plants"
                ],
                "summary": "Créer une nouvelle plante",
                "description": "Ajouter une nouvelle plante à la base de données",
                "operationId": "d005e00370ab9bd321d50ba5dde92042",
                "requestBody": {
                    "description": "Données de la plante à créer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "common_name",
                                    "watering_general_benchmark"
                                ],
                                "properties": {
                                    "common_name": {
                                        "description": "Nom commun de la plante",
                                        "type": "string",
                                        "example": "Rose"
                                    },
                                    "watering_general_benchmark": {
                                        "required": [
                                            "value",
                                            "unit"
                                        ],
                                        "properties": {
                                            "value": {
                                                "description": "Fréquence d'arrosage",
                                                "type": "string",
                                                "example": "7"
                                            },
                                            "unit": {
                                                "description": "Unité de temps",
                                                "type": "string",
                                                "example": "days"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plante créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "common_name": {
                                            "type": "string",
                                            "example": "Rose"
                                        },
                                        "watering_general_benchmark": {
                                            "properties": {
                                                "value": {
                                                    "type": "string",
                                                    "example": "7"
                                                },
                                                "unit": {
                                                    "type": "string",
                                                    "example": "days"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}